server/requirements.txt
fastapi
uvicorn
openai-whisper
edge-tts
numpy
torch
